backend-python/requirements.txt
aiofile==3.9.0
asyncpg==0.30.0
httpx==0.28.1
litestar[standard]==2.14.0
pydantic==2.10.6
pyjwt[crypto]==2.10.1
pyyaml==6.0.2
requests==2.32.3
SQLAlchemy==2.0.37
uvicorn==0.34.0
